+++ lib/ai_guardrails.dart
@@ -19,4 +19,6 @@ export 'src/scanners/repetition_scanner.dart';
 export 'src/scanners/url_scanner.dart';
 export 'src/scanners/language_scanner.dart';
 export 'src/scanners/code_execution_scanner.dart';
+export 'src/scanners/grounding_scanner.dart';
 export 'src/scanners/schema_validator.dart';
+export 'src/streaming_ai_guard.dart';